tofe_list split message functions to a separate file and changed fprintfs in tof_in.c to use these too
