+++ message.c
@@ -1,0 +1,27 @@
+#include <stdio.h>
+#include <stdarg.h>
+#include "message.h"
+
+const char *msg_level_str(tofe_list_msg_level level) {
+    switch(level) {
+        case TOFE_LIST_NORMAL:
+            return "";
+        case TOFE_LIST_INFO:
+            return "info";
+        case TOFE_LIST_WARNING:
+            return "warning";
+        case TOFE_LIST_ERROR:
+            return "error";
+        default:
+            return "";
+    }
+}
+
+void tofe_list_msg(tofe_list_msg_level level, const char *restrict format, ...) {
+    va_list ap;
+    va_start(ap, format);
+    fprintf(stderr, "tofe_list %s: ", msg_level_str(level));
+    vfprintf(stderr, format, ap);
+    fputc('\n', stderr);
+    va_end(ap);
+}
