+++ tofe_list.h
@@ -72,15 +72,7 @@ static const jibal_option cutfile_headers[] = {
         {0,                  0}
 };
 
-typedef enum tofe_list_msg_level {
-    TOFE_LIST_NORMAL = 0,
-    TOFE_LIST_INFO = 1,
-    TOFE_LIST_WARNING = 2,
-    TOFE_LIST_ERROR = 3
-} tofe_list_msg_level;
 
-const char *msg_level_str(tofe_list_msg_level level);
-void tofe_list_msg(tofe_list_msg_level level, const char *restrict format, ...);
 void tofe_files_free(list_files *files);
 int cutfile_parse_filename(jibal *jibal, cutfile *cutfile);
 int cutfile_parse_extensions(jibal *jibal, cutfile *cutfile, char **extensions, int n_ext);
